Add ADR instruction to load data segment variable into a register

+++ src/main/java/thb/fbi/leguan/instructions/DataTransferInstruction.java
@@ -20,7 +20,7 @@ public class DataTransferInstruction extends Instruction {
     public void simulate(InstructionArguments argument, PCRegister pc) {
         registerPaneController.clearFlagHighlighting();
         memoryController.clearMemoryHighlighting();
-        int dt_address = argument.getDt_Address();
+        long dt_address = argument.getDt_Address();
         Register Rn = argument.getRn();
         Register Rt = argument.getRt();
         this.dataTransferCode.simulate(dt_address, Rn, Rt);
