+++ src/main/java/thb/fbi/leguan/instructions/IDataTransferCode.java
@@ -6,5 +6,5 @@ import thb.fbi.leguan.simulation.Register;
  * Interface for definining and executing code of DataTransfer instructions
  */
 public interface IDataTransferCode {
-    public abstract void simulate(int dt_address, Register Rn, Register Rt);
+    public abstract void simulate(long dt_address, Register Rn, Register Rt);
 }
